requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.6.2=py310h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cython=3.0.8=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py310h06a4308_0
imageio=2.34.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.4=py310h06a4308_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.0.4=0
libcurand=10.3.6.39=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py310h5eee18b_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py310h06a4308_0
numpy=1.24.3=py310hd5efca6_0
numpy-base=1.24.3=py310h8e6c178_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_1
openssl=1.1.1w=h7f8727e_0
packaging=24.1=pypi_0
pillow=10.3.0=py310h5eee18b_0
pip=24.0=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.0=h12debd9_5
pytorch=2.2.0=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py310h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.2=py310h06a4308_0
scikit-image=0.24.0=pypi_0
scipy=1.14.0=pypi_0
setuptools=69.5.1=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py310h06a4308_0
tifffile=2024.7.2=pypi_0
tk=8.6.14=h39e8969_0
torchaudio=2.2.0=py310_cu121
torchtriton=2.2.0=py310
torchvision=0.17.0=py310_cu121
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=py310h06a4308_0
wheel=0.43.0=py310h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
